[Chore #159382843] Changed folder structure
