+++ .eslintrc
@@ -16,6 +16,7 @@
       "no-template-curly-in-string": 0,
       "import/newline-after-import": 0,
       "no-param-reassign": 0,
+      "import/no-mutable-exports": 0,
       "no-console":"off",
       "comma-dangle": 0,
       "curly": [
