+++ package.json
@@ -14,7 +14,7 @@
     "start": "cross-env NODE_ENV=development babel-node app.js",
     "dev-server": "cross-env NODE_ENV=development nodemon --exec babel-node app.js",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
-    "test": "cross-env NODE_ENV=test nyc mocha ./tests/**/'*.test.js' --recursive --timeout 50000 --exit || true",
+    "test": "cross-env process.env.NODE_ENV=test nyc mocha ./server/tests/**/'*.test.js' --recursive --timeout 50000 --exit || true",
     "test:watch": "npm test --watch",
     "lint": "eslint . --fix --ignore-path .gitignore ."
   },
