+++ server/helpers/validateEntry.js
