+++ server/tests/users.test.js
@@ -1,13 +1,13 @@
 import chai, { expect } from 'chai';
 import chaiHttp from 'chai-http';
-import server from '../app';
+import server from '../../app';
 
 chai.use(chaiHttp);
 
 describe('Users', () => {
   const user = {
     full_name: 'Samuel George',
-    email: 'sammy@sammy.com',
+    email: `${Math.random().toString(36).substring(2, 15)}@sammy.com`,
     password: 'password'
   };
 
@@ -18,6 +18,7 @@ describe('Users', () => {
       .end((err, res) => {
         expect(res.status).to.equal(201);
         expect(res.body).to.have.property('status').equal('success');
+        expect(res.body).to.have.property('users');
         expect(res.body).to.have.property('message').equal('Inserted one user');
         done();
       });
@@ -35,9 +36,9 @@ describe('Users', () => {
   });
   it('Should authenticate a valid user', (done) => {
     chai.request(server)
-      .post('/api/v1/auth/signin')
+      .post('/api/v1/auth/login')
       .send({
-        email: 'Tyler@mail.com',
+        email: 'sammy@sammy.com',
         password: 'password'
       })
       .end((err, res) => {
@@ -49,7 +50,7 @@ describe('Users', () => {
   });
   it('Should not signin a user with incomplete fields', (done) => {
     chai.request(server)
-      .post('/api/v1/auth/signin')
+      .post('/api/v1/auth/login')
       .send({ password: 'password' })
       .end((err, res) => {
         expect(res.status).to.equal(400);
@@ -60,15 +61,30 @@ describe('Users', () => {
   });
   it('Should not signin a user with incorrect password', (done) => {
     chai.request(server)
-      .post('/api/v1/auth/signin')
+      .post('/api/v1/auth/login')
       .send({
-        email: 'Tyler@mail.com',
+        email: 'admin@admin.com',
         password: 'passwordsss'
       })
       .end((err, res) => {
         expect(res.status).to.equal(400);
         expect(res.body).to.have.property('status').equal('Failed');
         expect(res.body).to.have.property('message').equal('Invalid email or password');
+        done();
+      });
+  });
+  it('Should not signup a user that exists', (done) => {
+    chai.request(server)
+      .post('/api/v1/auth/signup')
+      .send({
+        full_name: 'Samo alvin',
+        email: 'sammy@sammy.com',
+        password: 'passwordsss'
+      })
+      .end((err, res) => {
+        expect(res.status).to.equal(409);
+        expect(res.body).to.have.property('status').equal('Failed');
+        expect(res.body).to.have.property('message').equal('User already exists');
         done();
       });
   });
